internal/runtime/maps: use uintptr instead of uint32 for index in group
This avoids some zero-extension ops on 64-bit machines.

Based on khr@'s CL 619479.

For #54766.

+++ src/internal/runtime/maps/map.go
@@ -444,7 +444,7 @@ func (m *Map) getWithKeySmall(hash uintptr, key unsafe.Pointer) (unsafe.Pointer,
 	h2 := uint8(h2(hash))
 	ctrls := *g.ctrls()
 
-	for i := uint32(0); i < 8; i++ {
+	for i := uintptr(0); i < 8; i++ {
 		c := uint8(ctrls)
 		ctrls >>= 8
 		if c != h2 {
@@ -601,7 +601,7 @@ func (m *Map) growToTable() {
 		data: m.dirPtr,
 	}
 
-	for i := uint32(0); i < abi.SwissMapGroupSlots; i++ {
+	for i := uintptr(0); i < abi.SwissMapGroupSlots; i++ {
 		if (g.ctrls().get(i) & ctrlEmpty) == ctrlEmpty {
 			// Empty or deleted
 			continue
